Moved ApplicationConfiguration to TagzAppContext
