+++ src/TagzApp.Common/IApplicationConfigurationRepository.cs
@@ -1,0 +1,8 @@
+﻿// Ignore Spelling: Css
+
+namespace TagzApp.Common;
+
+public interface IApplicationConfigurationRepository
+{
+	Task SetValues(ApplicationConfiguration config);
+}
