+++ src/TagzApp.Web/Program.cs
@@ -13,6 +13,7 @@ public class Program
 {
 	private static void Main(string[] args)
 	{
+
 		var builder = WebApplication.CreateBuilder(args);
 
 		builder.Configuration.AddApplicationConfiguration();
@@ -21,7 +22,7 @@ public class Program
 		);
 
 		// Late bind the connection string so that any changes to the configuration made later on, or in the test fixture can be picked up.
-		builder.Services.AddDbContext<SecurityContext>();
+		builder.Services.AddSecurityContext(builder.Configuration);
 
 		builder.Services.AddDefaultIdentity<IdentityUser>(options =>
 				options.SignIn.RequireConfirmedAccount = true
@@ -79,7 +80,9 @@ public class Program
 			app.UseExceptionHandler("/Error");
 			// The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
 			app.UseHsts();
-		} else {
+		}
+		else
+		{
 			app.UseDeveloperExceptionPage();
 		}
