+++ src/TagzApp.WebTest/Fixtures/FixtureExtensions.cs
@@ -16,7 +16,9 @@ public static class FixtureExtensions
 		{
 			var testConfiguration = new Dictionary<string, string?>()
 			{
-				{ "ConnectionStrings:SecurityContextConnection", $"Data Source=TagzApp.Web.{id:N}.db" }
+				{ "ConnectionStrings:SecurityContextConnection", $"Data Source=TagzApp.Web.{id:N}.db" },
+				{ "ConnectionStrings:TagzAppSecurity", "" },
+				{ "ConnectionStrings:TagzApp", "" }
 			};
 			configuration.AddInMemoryCollection(testConfiguration);
 		});
