+++ src/TagzApp.WebTest/appsettings.test.json
@@ -6,7 +6,8 @@
 		}
 	},
 	"ConnectionStrings": {
-		"TagzAppSecurity": ""
+		"TagzAppSecurity": "",
+		"TagzApp": ""
 	},
 	"TestHostStartDelay": 0,
 	"ModerationEnabled": "false"
